requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
httpx==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
ngrok==1.4.0
npm==0.1.1
optional-django==0.1.0
paramiko==3.5.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.2
pytest-asyncio==0.23.7
PyNaCl==1.5.0
python-dotenv==1.1.0
requests==2.31.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3

# GUI Automation Dependencies
pip
pyyaml
opencv-python==4.10.0.84
pytesseract==0.3.13
pyautogui==0.9.54
pynput==1.7.7
python-xlib==0.33
pyperclip==1.9.0